Fix require.resolve and requiring non-bundled modules
Use resolve-from to resolve dependencies relative to the module's
directory. Resolve non-bundled dependencies before requiring them, which
actually makes it work.

Note that resolve-from itself cannot be bundled since it has
non-bundled dependencies. To load those dependencies they need to be
resolved, which requires resolve-from.

(That said, it only has built-in dependencies so those could be made
to skip resolve-from, if it's a performance concern.)